fix gas index calculation copy over json util from outdoors add brightness control for display update air gradient base

+++ AirGradient.h
@@ -132,7 +132,7 @@ typedef union {
 
 struct TMP_RH {
   float t;
-  float rh;
+  int rh;
   char t_char[10];
   char rh_char[10];
   TMP_RH_ErrorCode error;
